Fix signed/unsigned mismatch and other compiler warnings
- Fix C4018 signed/unsigned comparison warnings by changing loop variables
  to unsigned or adding explicit casts where appropriate
- Fix C4267 size_t to int/unsigned conversion warnings with explicit casts
- Fix C5033 by removing deprecated 'register' storage class specifiers
- Fix C4101 by removing unreferenced local variables
- Fix C4302/C4311/C4312 pointer truncation warnings using intptr_t

This resolves ~120 compiler warnings across:
- common/cmdlib.cpp, threads.cpp
- sdHLRAD/*.cpp (lightmap, qrad, sparse, trace, transfers, etc.)
- sdHLCSG/*.cpp (brush, qcsg, map)
- sdHLBSP/*.cpp (portals, solidbsp, surfaces, tjunc)
- sdHLVIS/*.cpp (flow, vis, zones)
- sdRIPENT/ripent.cpp

+++ src/sdhlt/common/cmdlib.cpp
@@ -209,7 +209,7 @@ void ExtractFile(const char* const path, char* dest)
 	int extension_pos, directory_pos;
 	getFilePositions(path,&extension_pos,&directory_pos);
 
-	int length = strlen(path);
+	int length = (int)strlen(path);
 
 	length -= directory_pos + 1;
 
@@ -221,7 +221,7 @@ void ExtractFileBase(const char* const path, char* dest)
 {
 	int extension_pos, directory_pos;
 	getFilePositions(path,&extension_pos,&directory_pos);
-	int length = extension_pos == -1 ? strlen(path) : extension_pos;
+	int length = extension_pos == -1 ? (int)strlen(path) : extension_pos;
 
 	length -= directory_pos + 1;
 
@@ -235,7 +235,7 @@ void ExtractFileExtension(const char* const path, char* dest)
 	getFilePositions(path,&extension_pos,&directory_pos);
 	if(extension_pos != -1)
 	{
-		int length = strlen(path) - extension_pos;
+		int length = (int)strlen(path) - extension_pos;
 	    memcpy(dest,path+extension_pos,length); //include extension '.'
 	    dest[length] = 0;
 	}
